Improves code safety and const correctness
Adds const qualifiers to buffer parameters in amod functions to prevent accidental modification.

Updates static arrays to be const to prevent modification.

Fixes potential buffer overflows in client command processing and display usage.

Adds null checks and paranoia checks to prevent crashes.

Fixes memory management issues with realloc by restoring counters on failure, preventing memory leaks.

Initializes buffers to prevent undefined behavior.

Updates data types for compatibility.

Addresses potential out-of-bounds access in sound effects.

style: add const to modder.c wrapper function parameters

+++ src/game/crash_handler_windows.c
@@ -15,7 +15,7 @@ void main_dump(FILE *fp)
 
 	fprintf(fp, "Main datadump:\n");
 
-	fprintf(fp, "game_options: %llu\n", game_options);
+	fprintf(fp, "game_options: %llu\n", (unsigned long long)game_options);
 	for (i = 0; i < 64; i++) {
 		tmp = 1llu << i;
 		if (game_options & tmp)
